Add files via upload
1: Create a basic home page with a header, navbar, sidebar, main and footer content. The contents of these are left to your creative imagination. As discussed, we may not be able to achieve the desired layout just by using the respective HTML tags. Instead, you will need to provide additional CSS styles to style them the way we intend them to be. The following resource on FlexBox can act as a starter to get the layout we desire: https://css-tricks.com/snippets/css/a-guide-to-flexbox/
 
2. Create a page, which displays any YouTube video as an iframe.
 
3. Create a page, which displays any Google Map location as an iframe.
 
4. Create a page, which has the ability to play audio and video. The audio and video files must reside in the project, and must not fetch them via the web.
 
Give appropriate styling to all of the pages above. Be cautious about what kind of CSS Selectors and combinators you use. 
Styling is left to your liking, so bring up your creative juices to make the best possible version.
 
It may be that depending on your current understanding and knowledge, you will have to read through or brush up on some concepts to complete the assignments. As always make use of your best friend - Google -  to achieve it.
 
Push the completed code to your GitHub repository and share the same either via the Google Form provided / Email.